convert to int

+++ DlgCreatePeer.cpp
@@ -196,7 +196,7 @@ bool DlgCreatePeer::check_configurations() {
     ui->lbl_err_disk->setStyleSheet("QLabel {color : red}");
     ui->lbl_err_disk->show();
     errors_exist = true;
-  } else if (disk.toInt() > Environment::Instance()->diskSize()) {
+  } else if (disk.toInt() > (int) Environment::Instance()->diskSize()) {
     ui->lbl_err_disk->setText(tr("Disk cannot be more than %1 GB.").arg(Environment::Instance()->diskSize()));
     ui->lbl_err_disk->setStyleSheet("QLabel {color : red}");
     ui->lbl_err_disk->show();
